Moved implementation from the manager to the editor subsystem
