+++ Source/NamingConventionValidation/Public/NamingConventionValidationManager.h
@@ -2,6 +2,7 @@
 
 #include "AssetData.h"
 #include "CoreMinimal.h"
+#include "EditorNamingValidatorSubsystem.h"
 #include "Engine/EngineTypes.h"
 #include "UObject/Object.h"
 
@@ -11,14 +12,6 @@ class FLogWindowManager;
 class SLogWindow;
 class SWindow;
 class UUnitTest;
-
-enum class ENamingConventionValidationResult : uint8
-{
-    Invalid,
-    Valid,
-    Unknown,
-    Excluded
-};
 
 UCLASS()
 class NAMINGCONVENTIONVALIDATION_API UNamingConventionValidationManager final : public UObject
@@ -33,8 +26,8 @@ public:
     virtual ~UNamingConventionValidationManager();
 
     ENamingConventionValidationResult IsAssetNamedCorrectly( const FAssetData & asset_data, FText & error_message ) const;
-    int32 ValidateAssets( const TArray< FAssetData > & asset_data_list, bool skip_excluded_directories = true, bool show_if_no_failures = true ) const;
     void ValidateOnSave( const TArray< FAssetData > & asset_data_list ) const;
+    int32 ValidateAssets( const TArray< FAssetData > & asset_data_list, bool skip_excluded_directories = true, bool show_if_no_failures = true ) const;
     void ValidateSavedPackage( const FName package_name );
 
 protected:
@@ -44,5 +37,4 @@ protected:
 
 private:
     ENamingConventionValidationResult DoesAssetMatchNameConvention( const FAssetData & asset_data, const FName asset_class, FText & error_message ) const;
-    void GetRenamedAssetSoftObjectPath( FSoftObjectPath & renamed_soft_object_path, const FAssetData & asset_data ) const;
 };
